+++ reactive-banana/CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog for the `reactive-banana** package
 
 **unreleased**
 
-* Add `mergeWith` combinator. [#163][], [#220][]
+* Add `merge` and `mergeWith` combinators. [#163][], [#220][]
 * Make internal SCC pragmas compatible with the GHC 9.0 parser. [#208][]
 
   [#163]: https://github.com/HeinrichApfelmus/reactive-banana/pull/163